remove superfluous suffix in filenames
