+++ docs/demos/comments.mdx
@@ -4,7 +4,7 @@ menu: Demos
 route: demos/comments
 ---
 
-import DemoComments from './src/demo-comments';
+import CommentsDemo from './src/comments';
 
 # Comments
 
@@ -24,4 +24,4 @@ You can improve this demo by formally building out a commenting workflow with re
 - Annotation `data-` attributes.
 - Custom CSS.
 
-<DemoComments />
+<CommentsDemo />
