+++ docs/demos/diff.mdx
@@ -4,7 +4,7 @@ menu: Demos
 route: demos/diff
 ---
 
-import DemoDiff from './src/demo-diff';
+import DiffDemo from './src/diff';
 
 # Diff
 
@@ -31,4 +31,4 @@ It would be interesting to find a way to apply Github diff styles that highlight
 - Constructing annotations from other data.
 - Custom CSS.
 
-<DemoDiff />
+<DiffDemo />
