+++ docs/demos/ebook.mdx
@@ -4,7 +4,7 @@ menu: Demos
 route: demos/ebook
 ---
 
-import DemoEbook from './src/demo-ebook';
+import EbookDemo from './src/ebook';
 
 # Ebook
 
@@ -26,4 +26,4 @@ You can create more interactive bookmark workflows by taking advantage of the mo
 - Customizable annotations.
 - Permalinks to annotated nodes.
 
-<DemoEbook />
+<EbookDemo />
