+++ docs/demos/labeling.mdx
@@ -4,7 +4,7 @@ menu: Demos
 route: demos/labeling
 ---
 
-import DemoLabeling from './src/demo-labeling';
+import LabelingDemo from './src/labeling';
 
 # Labeling
 
@@ -21,4 +21,4 @@ A better machine learning pipeline that returns relevant annotations would make 
 - Custom post-annotation rendering with web technologies.
 - Transforming/replacing annotated nodes text content.
 
-<DemoLabeling />
+<LabelingDemo />
