+++ docs/demos/search.mdx
@@ -4,7 +4,7 @@ menu: Demos
 route: demos/search
 ---
 
-import DemoSearch from './src/demo-search';
+import SearchDemo from './src/search';
 
 # Search
 
@@ -20,4 +20,4 @@ You can use formal search libraries or integrate with a backend search API.  Not
 ## Concepts
 - Permalinks to annotated nodes.
 
-<DemoSearch />
+<SearchDemo />
