+++ docs/packages/unified-doc-react/annotations.mdx
@@ -7,7 +7,7 @@ route: packages/unified-doc-react/annotations
 import { Link } from 'gatsby';
 import { useState } from 'react';
 
-import ExampleAnnotations from './src/example-annotations';
+import ExampleAnnotations from './src/annotations';
 
 # Annotations
