+++ docs/packages/unified-doc-react/content.mdx
@@ -4,7 +4,7 @@ menu: unified-doc-react
 route: packages/unified-doc-react/content
 ---
 
-import ExampleContent from './src/example-content';
+import ExampleContent from './src/content';
 
 # Content
