+++ docs/packages/unified-doc-react/managing-annotations.mdx
@@ -4,7 +4,7 @@ menu: unified-doc-react
 route: packages/unified-doc-react/managing-annotations
 ---
 
-import ExampleManagingAnnotations from './src/example-managing-annotations';
+import ExampleManagingAnnotations from './src/managing-annotations';
 
 # Managing Annotations
