+++ docs/packages/unified-doc-react/plugins.mdx
@@ -4,7 +4,7 @@ menu: unified-doc-react
 route: packages/unified-doc-react/plugins
 ---
 
-import ExamplePlugins from './src/example-plugins';
+import ExamplePlugins from './src/plugins';
 
 # Plugins
