+++ docs/packages/unified-doc-react/styling.mdx
@@ -4,7 +4,7 @@ menu: unified-doc-react
 route: packages/unified-doc-react/styling
 ---
 
-import ExampleStyling from './src/example-styling';
+import ExampleStyling from './src/styling';
 
 # Styling
